# docs/api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "AURA Care-Chain API",
        "description": "API REST do AURA Care-Chain \u2014 assistente de sa\u00fade domiciliar voice-first\npara idosos com Parkinson, com cadeia log\u00edstica de seguran\u00e7a da casa.\n\nConsumida por tr\u00eas clientes: app Flutter (Android/iOS/Web),\napp React Native e painel administrativo Angular.\n\nRegras de ouro: nunca prescreve nem diagnostica, o escore \u00e9 sempre\nexplic\u00e1vel (fatores + pesos) e nenhum dado de sa\u00fade \u00e9 gravado sem\nconsentimento LGPD.\n",
        "contact": {
            "name": "Equipe AURA \u2014 FIAP Smart HAS"
        },
        "license": {
            "name": "Uso acad\u00eamico \u2014 FIAP 2026"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/signals": {
            "post": {
                "tags": [
                    "3. Monitoramento 360"
                ],
                "summary": "Registra um sinal (voz, auto-relato, uso do app ou wearable)",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSignalRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignalCreatedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Consentimento LGPD n\u00e3o aceito (RN-001)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "CONSENT_REQUIRED": {
                                        "value": {
                                            "error": {
                                                "code": "CONSENT_REQUIRED",
                                                "message": "Aceite a Pol\u00edtica de Privacidade antes de registrar dados de sa\u00fade.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scores/recompute": {
            "post": {
                "tags": [
                    "3. Monitoramento 360"
                ],
                "summary": "Recalcula o escore; sem dimens\u00e3o, devolve a de maior risco",
                "operationId": "recompute",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecomputeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScoreResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Consentimento LGPD n\u00e3o aceito (RN-001)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "CONSENT_REQUIRED": {
                                        "value": {
                                            "error": {
                                                "code": "CONSENT_REQUIRED",
                                                "message": "Aceite a Pol\u00edtica de Privacidade antes de registrar dados de sa\u00fade.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dimens\u00e3o fora da configura\u00e7\u00e3o do escore",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNKNOWN_DIMENSION": {
                                        "value": {
                                            "error": {
                                                "code": "UNKNOWN_DIMENSION",
                                                "message": "Dimens\u00e3o desconhecida: telepatia",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/homes/{homeId}/signals": {
            "get": {
                "tags": [
                    "3. Monitoramento 360"
                ],
                "summary": "Hist\u00f3rico de sinais da casa, com filtros e pagina\u00e7\u00e3o",
                "operationId": "list_1",
                "parameters": [
                    {
                        "name": "homeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "mobility",
                                "sleep",
                                "cognition",
                                "mood",
                                "environment",
                                "adherence",
                                "vitals"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SignalResponse"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/homes/{homeId}/scores": {
            "get": {
                "tags": [
                    "3. Monitoramento 360"
                ],
                "summary": "Hist\u00f3rico de escores da casa",
                "operationId": "history",
                "parameters": [
                    {
                        "name": "homeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScoreResponse"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateSignalRequest": {
                "required": [
                    "homeId",
                    "source",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "homeId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "mobility",
                            "sleep",
                            "cognition",
                            "mood",
                            "environment",
                            "adherence",
                            "vitals"
                        ]
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "voice",
                            "self_report",
                            "usage",
                            "wearable"
                        ]
                    },
                    "value": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object",
                            "example": {
                                "event": "near_fall",
                                "place": "bathroom"
                            }
                        },
                        "example": {
                            "event": "near_fall",
                            "place": "bathroom"
                        }
                    }
                }
            },
            "SignalCreatedResponse": {
                "type": "object",
                "properties": {
                    "signalId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "RecomputeRequest": {
                "required": [
                    "homeId"
                ],
                "type": "object",
                "properties": {
                    "homeId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "dimension": {
                        "type": "string",
                        "description": "Omitido = recalcula todas e devolve a de maior risco",
                        "example": "mobility"
                    }
                }
            },
...
            "SignalResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "mobility",
                            "sleep",
                            "cognition",
                            "mood",
                            "environment",
                            "adherence",
                            "vitals"
                        ]
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "voice",
                            "self_report",
                            "usage",
                            "wearable"
                        ]
                    },
                    "value": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    },
                    "capturedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/Body"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}